% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.r
\docType{data}
\name{acetate}
\alias{acetate}
\alias{materials}
\alias{acrylic}
\alias{polycarbonate}
\alias{plexiglas}
\alias{polystyrene}
\alias{polyester}
\alias{polyvynil_chloride}
\alias{optical_glass}
\alias{plastic_film}
\alias{plastic_sheet}
\title{Spectral data for filters of different 'materials'}
\format{A vector of character strings.}
\usage{
acetate

acrylic

polycarbonate

plexiglas

polystyrene

polyester

polyvynil_chloride

optical_glass

plastic_film

plastic_sheet
}
\description{
Names of datasets containing the wavelengths and tabulated values for
fractional spectral trabsmittance according to the material they are
made off.
}
\examples{
acrylic

# select short-pass filters
filters.mspct[acrylic]

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
