% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-accesors.r
\docType{data}
\name{all_filter_accessors}
\alias{all_filter_accessors}
\title{Names of different accessors}
\format{
A vector of character strings.
}
\usage{
all_filter_accessors
}
\description{
The package makes available a collection of character vectors that can be
used to extract spectral data for specific subsets of the filters. Some
correspond to supplier names, other to materials, types, colours and
typical uses. The data object described here lists the names of all these
accesors, as a sort of index to this collection of data.
}
\examples{
all_filter_accessors

}
\keyword{datasets}
