% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-selectors.r
\docType{data}
\name{all_filter_selectors}
\alias{all_filter_selectors}
\alias{selectors}
\alias{all_screen_net_selectors}
\title{Names of different indexing vectors}
\format{
A vector of character strings.

An object of class \code{character} of length 7.
}
\usage{
all_filter_selectors

all_screen_net_selectors
}
\description{
The package makes available character vectors that can be used to extract
spectral data for specific subsets of the filters, screens, etc. Some
correspond to supplier names, others to materials, types, colours and typical
uses. The data objects described here list the names of all these vectors
of names grouped by the collection of spectra they index.
}
\examples{
all_screen_net_selectors
all_filter_selectors

}
\seealso{
\code{\link{filters.mspct}} and \code{\link{screens_nets.mspct}}
}
\keyword{datasets}
