% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.r
\docType{data}
\name{band_pass}
\alias{band_pass}
\alias{types}
\alias{long_pass}
\alias{short_pass}
\title{Spectral data for filters of different 'types'}
\format{A vector of character strings.}
\usage{
band_pass

long_pass

short_pass
}
\description{
Names of datasets containing the wavelengths and tabulated values for
fractional spectral trabsmittance according to their type. Band-pass
filters transmit a range of wavelengths, but block shorter and longer
ones. Long-pass filters absorb wavelengths shorter than their cut-off and
short-pass filters absorb wavelengths longer than their cut-off.
}
\note{
These vectors include only optical-glass filters with well defined
characteristics.
}
\examples{
band_pass

# select band-pass filters
filters.mspct[band_pass]

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
