% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpi-visqueen.r
\docType{data}
\name{bpi_visqueen_filters}
\alias{bpi_visqueen_filters}
\title{Indexing of greenhouse films from BPI Visqueen}
\format{
A vector of character strings.
}
\usage{
bpi_visqueen_filters
}
\description{
Vector of names of member objects containing spectral transmittance data for different films used as
greenhouse cladding supplied by BPI Visqueen.
}
\note{
Individual measurements of single samples, not a specification.
  Instrument used: Shimadzu UV-2501 PC (Shimadzu Suzhou Instruments
  Manufacturing Co. Ltd, China) equipped with an integrating sphere. Some of
  these film types are no longer produced.
}
\examples{
bpi_visqueen_filters

}
\references{
\url{https://www.rpc-bpi.com/sector/horticultural-solutions/}
}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
