% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{clear_filters}
\alias{clear_filters}
\alias{colors}
\alias{neutral_filters}
\alias{uv_filters}
\alias{blue_filters}
\alias{blue_green_filters}
\alias{green_filters}
\alias{yellow_filters}
\alias{orange_filters}
\alias{red_nir_filters}
\alias{heat_filters}
\title{Spectral data for filters of different 'colors'}
\format{A vector of character strings.}
\usage{
clear_filters

neutral_filters

uv_filters

blue_filters

blue_green_filters

green_filters

yellow_filters

orange_filters

red_nir_filters

heat_filters
}
\description{
Names of datasets containing the wavelengths and tabulated values for
fractional spectral trabsmittance for filters of given colours.
}
\examples{
clear_filters

# select filters of amber, yellow, orange color ("blue absorbing")
filters.mspct[yellow_filters]

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
