% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etola.r
\docType{data}
\name{etola_filters}
\alias{etola_filters}
\title{Indexing of PE films supplied by Etola Oy}
\format{
A vector of character strings.
}
\usage{
etola_filters
}
\description{
Vectors of names of member objects containing spectral transmittance data for polythene film supplied by
Etola Oy (Finland).
}
\note{
Individual measurements of single samples, not a specification.
}
\examples{
etola_filters

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
