% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.r
\docType{data}
\name{acetate_filters}
\alias{acetate_filters}
\alias{acrylic_filters}
\alias{polycarbonate_filters}
\alias{plexiglas_filters}
\alias{polystyrene_filters}
\alias{polyester_filters}
\alias{polyvynil_chloride_filters}
\alias{optical_glass_filters}
\alias{plastic_film_filters}
\alias{plastic_films}
\alias{plastic_sheet_filters}
\alias{plastic_sheets}
\alias{plastic_dome_filters}
\alias{plastic_domes}
\title{Indexing of filters made from different materials}
\format{
A vector of character strings.

An object of class \code{character} of length 9.

An object of class \code{character} of length 3.

An object of class \code{character} of length 9.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 1.

An object of class \code{character} of length 68.

An object of class \code{character} of length 31.

An object of class \code{character} of length 31.

An object of class \code{character} of length 14.

An object of class \code{character} of length 14.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
data(acetate_filters)

data(acrylic_filters)

data(polycarbonate_filters)

data(plexiglas_filters)

data(polystyrene_filters)

data(polyester_filters)

data(polyvynil_chloride_filters)

data(optical_glass_filters)

data(plastic_film_filters)

data(plastic_films)

data(plastic_sheet_filters)

data(plastic_sheets)

data(plastic_dome_filters)

data(plastic_domes)
}
\description{
Vectors of names of member objects containing fractional spectral
transmittance data for films (very flexible) and sheets (rigid or nearly
rigid) grouped by the material they are made of.
}
\details{
The vectors can be used to extract subsets of spectra from the
 collection in \code{filter.mspct}. Some spectra are included in more
 than one vector.
}
\examples{
acrylic_filters

# extract spectral data for filters made from acrylic
filters.mspct[acrylic_filters]

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
