% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee.data.r
\docType{data}
\name{lee}
\alias{lee}
\title{Filter spectra data for Lee theatrical filter}
\format{A vector of character strings.}
\usage{
lee
}
\description{
Datasets containing the wavelengths at a 1 nm interval and
tabulated values spectral transmittance for different filters.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}
}
\examples{
lee

}
\references{
\url{http://www.leefilters.com/lighting/colour-details.html}
}
\keyword{datasets}
