% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdermit.r
\docType{data}
\name{mcdermit_filters}
\alias{mcdermit_filters}
\title{Indexing of McDermit "polyester" films}
\format{
A vector of character strings.
}
\usage{
mcdermit_filters
}
\description{
Vector of names of member objects containing spectral transmittance data for
plastic films produced by McDermit Autotype.
}
\note{
Instrument used: Shimadzu UV-2501 PC (Shimadzu Suzhou Instruments
  Manufacturing Co. Ltd, China) equipped with an integrating sphere.
}
\examples{
mcdermit_filters

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
