% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midopt.r
\docType{data}
\name{midopt}
\alias{midopt}
\title{Total transmittance for MidOpt filters}
\format{A vector of character strings.}
\usage{
midopt
}
\description{
Internal transmittance for MidOpt machine-vision optical glass, acrylic and
interference filters. Data extracted from MidOpt's pdf data files with
help of R package 'tabulizer'. MidOpt is a trade
name of Midwest Optical Systems, Palatine, IL 60067 USA
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}
}
\note{
Data downloaded from \url{http://midopt.com/filters/} on 2017-12-26.
Republished with permission.
Copyright (c) Midwest Optical Systems, Inc. (MidOpt)
\url{http://www.midopt.com/}
}
\section{DISCLAIMER from MidOpt}{
 data are typical values and may vary by
plus/minus 10 nm.
}

\examples{
midopt

}
\keyword{datasets}
