% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.dishes.r
\docType{data}
\name{petri_dishes}
\alias{petri_dishes}
\title{Total transmittance for  Petri dishes}
\format{A vector of character strings.}
\usage{
petri_dishes
}
\description{
Total transmittance for the covers of Petri dishes. Data are to be used as
"reference values only" as values may change in time and production batch.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}

\url{https://www.thermofisher.com/}
}
\note{
Data measured with an Agilent 8453 spectrophotometer (based on diode array).

Idependent measurement of a single sample, not a specification.
}
\examples{
petri_dishes

}
\keyword{datasets}
