% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosco.r
\docType{data}
\name{rosco}
\alias{rosco}
\title{Filter spectra data for Rosco thetrical filters or 'gels'}
\format{A vector of character strings.}
\usage{
rosco
}
\description{
Datasets containing the wavelengths at a 1 nm interval and
tabulated values spectral transmittance for different filters.
Unused (new).
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}

\url{http://www.rosco.com/filters/index.cfm}
}
\note{
Instrument used: Shimadzu UV-2501 PC (Shimadzu Suzhou Instruments Manufacturing
Co. Ltd, China) equiped with an integrating sphere.

Idependent measurement of a single sample, not a specification.
}
\examples{
rosco

}
\keyword{datasets}
