% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screens.R
\docType{data}
\name{arrigoni_screens}
\alias{arrigoni_screens}
\alias{criadolopez_screens}
\alias{howitec_screens}
\alias{huachangyarns_screens}
\alias{jiangsuhuachangyarns_screens}
\alias{mallastextiles_screens}
\alias{oerlemansplastics_screens}
\alias{svensson_screens}
\title{Spectral data for filters made from different materials}
\format{
A vector of character strings.

An object of class \code{character} of length 17.

An object of class \code{character} of length 40.

An object of class \code{character} of length 5.

An object of class \code{character} of length 10.

An object of class \code{character} of length 45.

An object of class \code{character} of length 3.

An object of class \code{character} of length 32.
}
\usage{
arrigoni_screens

criadolopez_screens

howitec_screens

huachangyarns_screens

jiangsuhuachangyarns_screens

mallastextiles_screens

oerlemansplastics_screens

svensson_screens
}
\description{
Names of datasets containing the wavelengths and tabulated values for
fractional spectral transmittance grouped by supplier.
}
\examples{
arrigoni_screens

# select filters supplied by Arrigoni
screens.mspct[arrigoni_screens[1:5]]

}
\seealso{
\code{\link{screens.mspct}}
}
\keyword{datasets}
