% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqg.r
\docType{data}
\name{uqg_filters}
\alias{uqg_filters}
\title{Total transmittance for UQG filters}
\format{
A vector of character strings.
}
\usage{
uqg_filters
}
\description{
Total transmittance for UQG optical glass absorptive and interference
(dichroic) filters. Data measured by P. J. Aphalo with an Agilet 8453 array
spectrophotometer. Some of the filters are Schott glass and we have retained
the Schott type names.
}
\details{
The variables are as follows:

\itemize{ \item w.length (nm) \item Tfr (fraction) }
}
\note{
Filter supplier: UQG Ltd. The Norman Industrial Estate, Milton,
  Cambridge, England. \url{https://www.uqgoptics.com/}
}
\section{DISCLAIMER}{
 These spectral transmittance data are based on the
  measurement of a single filter of each type. Filters from other production
  batches may differ in their spectral properties, specially for wavelengths
  outside the range described in the manufacturer's published specifications.
  The characterisitics of some types of glass filters may change with time
  through surface oxidation or by effect of exposure to radiation including
  solar radiation. The data provided is to be used only as a coarse guide.
  When filter properties are important the filter actually used should be
  measured.
}

\examples{
uqg_filters

}
\keyword{datasets}
