% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl-horticulture.r
\docType{data}
\name{xl_horticulture_filters}
\alias{xl_horticulture_filters}
\title{Filter spectra data for commercial grenhouse films from XL Horticulture}
\format{
A vector of character strings.
}
\usage{
xl_horticulture_filters
}
\description{
Datasets containing the wavelengths at a 1 nm interval and
tabulated values spectral transmittance for different films used as
greenhouse cladding.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}

\url{https://www.xlhorticulture.co.uk/}

Instrument used: Shimadzu UV-2501 PC (Shimadzu Suzhou Instruments Manufacturing
Co. Ltd, China) equiped with an integrating sphere.
}
\note{
Idependent measurement of a single sample, not a specification.
}
\examples{
xl_horticulture_filters

}
\keyword{datasets}
