% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-selectors.r
\docType{data}
\name{all_filter_selectors}
\alias{all_filter_selectors}
\alias{selectors}
\alias{all_screen_selectors}
\title{Names of different selectors}
\format{
A vector of character strings.

An object of class \code{character} of length 8.
}
\usage{
all_filter_selectors

all_screen_selectors
}
\description{
The package makes available character vectors that can be used to extract
spectral data for specific subsets of the filters, screens, etc. Some
correspond to supplier names, other to materials, types, colours and typical
uses. The data objects described here list the names of all these vectors
of names corresponding to the different collections of spectra..
}
\examples{
all_screen_selectors

}
\keyword{datasets}
