% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpi-visqueen.r
\docType{data}
\name{bpi_visqueen_filters}
\alias{bpi_visqueen_filters}
\title{Filter spectra data for commercial grenhouse films from BPI Visqueen}
\format{
A vector of character strings.
}
\usage{
bpi_visqueen_filters
}
\description{
Datasets containing the wavelengths at a 1 nm interval and
tabulated values spectral transmittance for different films used as
greenhouse cladding.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}

\url{https://www.rpc-bpi.com/sector/horticultural-solutions/}

Instrument used: Shimadzu UV-2501 PC (Shimadzu Suzhou Instruments Manufacturing
Co. Ltd, China) equiped with an integrating sphere.
}
\note{
Idependent measurement of a single sample, not a specification.
}
\examples{
bpi_visqueen_filters

}
\keyword{datasets}
