% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{clear_filters}
\alias{clear_filters}
\alias{colors}
\alias{neutral_filters}
\alias{uv_filters}
\alias{blue_filters}
\alias{blue_green_filters}
\alias{green_filters}
\alias{yellow_filters}
\alias{orange_filters}
\alias{red_nir_filters}
\alias{heat_filters}
\alias{uvir_cut_filters}
\title{Spectral data for filters of different 'colors'}
\format{
A vector of character strings.

An object of class \code{character} of length 30.

An object of class \code{character} of length 44.

An object of class \code{character} of length 9.

An object of class \code{character} of length 22.

An object of class \code{character} of length 6.

An object of class \code{character} of length 14.

An object of class \code{character} of length 8.

An object of class \code{character} of length 24.

An object of class \code{character} of length 4.

An object of class \code{character} of length 7.
}
\usage{
clear_filters

neutral_filters

uv_filters

blue_filters

blue_green_filters

green_filters

yellow_filters

orange_filters

red_nir_filters

heat_filters

uvir_cut_filters
}
\description{
Names of datasets containing the wavelengths and tabulated values for
fractional spectral trabsmittance for filters of given colours.
}
\examples{
clear_filters

# select filters of amber, yellow, orange color ("blue absorbing")
filters.mspct[yellow_filters]

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
