% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courtaulds.r
\docType{data}
\name{courtaulds_filters}
\alias{courtaulds_filters}
\title{Filter spectra data for cellulose di-acetate film}
\format{
A vector of character strings.
}
\usage{
courtaulds_filters
}
\description{
Datasets containing the wavelengths at a 1 nm interval and
tabulated values spectral transmittance for Courtaulds di-acetate films.

1) Films (0.115 mm thick) stretched 25 cm below lamps Q-Panel UVB313 and
exposed to UV radiaition from different lengths of time per day. All films
were installed and sampled at the same time. The digits give the number of
minutes per day that the lamps were on.

2) New films of different thicknesses.
}
\details{
For each spectrum in the collection the variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}
}
\note{
Instrument used: Hewlett-Packard 8453 diode array spectrophotometer
  without an integrating sphere.

Idependent measurement of a single sample, not a specification.
}
\examples{
courtaulds_filters

}
\keyword{datasets}
