% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etola.r
\docType{data}
\name{etola_filters}
\alias{etola_filters}
\title{PE films supplied by Etola Oy}
\format{
A vector of character strings.
}
\usage{
etola_filters
}
\description{
Dataset containing the wavelengths at a 1 nm interval and
tabulated values spectral transmittance for different filters supplied by
Etola Oy (Finland).
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}
}
\note{
Idependent measurement of a single sample, not a specification.
}
\examples{
etola_filters

}
\keyword{datasets}
