% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters-mspct.r
\docType{data}
\name{filters.mspct}
\alias{filters.mspct}
\title{Spectral irradiance for diverse optical filters}
\format{
A \code{"filter_mspct"} object containing several
  \code{"filter_spct"}.
}
\usage{
filters.mspct
}
\description{
A collection of transmittance spectra for optical filters
  from different suppliers.
}
\details{
The \code{"filter_mspct"} object contains \code{"filter_spct"}
  objects with spectral transmittance data.

  The variables in each member spectrum are as follows: \itemize{ \item
  w.length (nm) \item Tfr (transmittance expressed as a fraction of one) }
}
\note{
Please see the help corresponding to each supplier for details.
}
\examples{
filters.mspct$Schott_UG11
filters.mspct[[1]]
names(filters.mspct)

}
\keyword{datasets}
