% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foiltek.r
\docType{data}
\name{foiltek_filters}
\alias{foiltek_filters}
\title{Spectral Transmittance for Plastic Sheets Supplied by Foiltek}
\format{
A vector of character strings.
}
\usage{
foiltek_filters
}
\description{
Total transmittance for samples of clear plastic sheets. PC_UV is 3 mm thick
Makrolife sheet manufactured by Arlaplast Ab (Sweden). PC is generic 3 mm
thick polycarbonate, PS is 3 mm thick polystyrene PET_G is polyester approx 3
mm thick and PVC is polyvinylchloride approximately 1.5--2 mm thick all from
unknown manufacturers but stock product in Foiltek's catalogue. Data are to
be used as "reference values only" as values may change in time and
production batch.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}

Sample chips supplied free of charge by Foiltek on 02.07.2014
\url{https://foiltek.fi/}
}
\note{
Data measured with an Agilent 8453 spectrophotometer (based on diode
  array).

Idependent measurement of a single sample, not a specification.
}
\examples{
foiltek_filters

}
\keyword{datasets}
