% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdermit.r
\docType{data}
\name{mcdermit_filters}
\alias{mcdermit_filters}
\title{Filter spectra data for McDermit Autostat CT5 "polyester" filter}
\format{
A vector of character strings.
}
\usage{
mcdermit_filters
}
\description{
Datasets containing the wavelengths at a 1 nm interval and
tabulated values spectral transmittance for filter.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}
\url{http://autotype.macdermid.com/}
}
\note{
Instrument used: Shimadzu UV-2501 PC (Shimadzu Suzhou Instruments Manufacturing
Co. Ltd, China) equiped with an integrating sphere.
}
\examples{
mcdermit_filters

}
\keyword{datasets}
