% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petri.dishes.r
\docType{data}
\name{petri_dishes}
\alias{petri_dishes}
\title{Indexing of Petri dishes}
\format{
A vector of character strings.
}
\usage{
petri_dishes
}
\description{
Vectors of names of member objects with spectral data for the covers of Petri
dishes. Data are to be used as "reference values only" as values may change
in time and with production batch.
}
\note{
Data measured with an Agilent 8453 spectrophotometer (based on diode
  array). Supplier of the plastic Petri dishes: Thermo Fisher Scientific.

Individual measurement of a single sample, not a specification.
}
\examples{
petri_dishes

filters.mspct[petri_dishes]

}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
