% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photography-filters.r
\docType{data}
\name{photography_filters}
\alias{photography_filters}
\alias{hoya_filters}
\alias{firecrest_filters}
\alias{haida_filters}
\alias{zomei_filters}
\alias{fake_unbranded_filters}
\alias{baader_filters}
\alias{uvroptics_filters}
\alias{tiffen_filters}
\alias{bw_filters}
\alias{heliopan_filters}
\alias{zeiss_filters}
\alias{kenko_filters}
\alias{rocolax_filters}
\alias{kolarivision_filters}
\alias{fotga_filters}
\alias{tangsinuo_filters}
\alias{purshee_filters}
\alias{knightx_filters}
\title{Total transmittance for photography filters}
\format{
A vector of character strings.

An object of class \code{character} of length 11.

An object of class \code{character} of length 7.

An object of class \code{character} of length 2.

An object of class \code{character} of length 10.

An object of class \code{character} of length 2.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 3.

An object of class \code{character} of length 1.

An object of class \code{character} of length 7.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 3.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 16.

An object of class \code{character} of length 2.

An object of class \code{character} of length 4.
}
\usage{
photography_filters

hoya_filters

firecrest_filters

haida_filters

zomei_filters

fake_unbranded_filters

baader_filters

uvroptics_filters

tiffen_filters

bw_filters

heliopan_filters

zeiss_filters

kenko_filters

rocolax_filters

kolarivision_filters

fotga_filters

tangsinuo_filters

purshee_filters

knightx_filters
}
\description{
Total transmittance for filters intended for use on camera lenses from
various brands and suppliers: B+W, Formatt-Hitech (Hitech, Firecrest), Fotga,
Haida, Heliopan, Hoya, KinghtX, KolariVision, Purshee, Rocolax, Tangsinuo,
Tiffen, UVRoptics, Zeiss, and Zomei. Data are to be used as \emph{"reference
values only"} as values may change in time and production batch or glass
melt. The fine structure of the spectra depends on the wavelength resolution
of the spectrometer used and post-processing applied to the data. In some
cases spectra for multiple filters of the same brand are type are included
and show such variation.
}
\details{
The variables are as follows:

\itemize{ \item w.length (nm) \item Tfr (fraction) }
}
\note{
Data measured with an Agilent 8453 spectrophotometer (based on diode
  array).

Idependent measurement of a single sample, not a specification.
}
\section{Supplier's web sites, if known}{

  \url{https://www.baader-planetarium.com/en/},
  \url{https://www.formatt-hitech.com/}, \url{https://www.fotga.com/},
  \url{https://www.haidaphoto.com/en/}, \url{https://www.heliopan.de/},
  \url{https://hoyafilter.com/}, \url{https://kenkoglobal.com/},
  \url{https://knightx-photography.aliexpress.com/store/924165},
  \url{https://kolarivision.com/},
  \url{https://purshee.aliexpress.com/store/1996249},
  \url{https://schneiderkreuznach.com/en/photo-optics},
  \url{https://tangsinuo.aliexpress.com/store/4314008},
  \url{https://tiffen.com/}, \url{https://www.uvroptics.com/},
  \url{https://www.zomei.com/}.
}

\examples{
photography_filters

}
\keyword{datasets}
