% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photography-filters.r
\docType{data}
\name{photography_filters}
\alias{photography_filters}
\alias{hoya_filters}
\alias{firecrest_filters}
\alias{haida_filters}
\alias{zomei_filters}
\alias{fake_unbranded_filters}
\alias{baader_filters}
\alias{uvroptics_filters}
\alias{tiffen_filters}
\alias{bw_filters}
\alias{heliopan_filters}
\alias{zeiss_filters}
\alias{kenko_filters}
\alias{rocolax_filters}
\alias{kolarivision_filters}
\alias{fotga_filters}
\title{Total transmittance for photography filters}
\format{
A vector of character strings.

An object of class \code{character} of length 11.

An object of class \code{character} of length 7.

An object of class \code{character} of length 1.

An object of class \code{character} of length 10.

An object of class \code{character} of length 5.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 7.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 3.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
photography_filters

hoya_filters

firecrest_filters

haida_filters

zomei_filters

fake_unbranded_filters

baader_filters

uvroptics_filters

tiffen_filters

bw_filters

heliopan_filters

zeiss_filters

kenko_filters

rocolax_filters

kolarivision_filters

fotga_filters
}
\description{
Total transmittance for filters intended for use on camera lenses from
various suppliers: Formatt Hitech, Haida, Heliopan, Hoya, Rocolax and Zomei.
Data are to be used as "reference values only" as values may change in time
and production batch.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item Tfr (fraction)
}
}
\note{
Data measured with an Agilent 8453 spectrophotometer (based on diode array).

Idependent measurement of a single sample, not a specification.
}
\section{Supplier's web sites}{

\url{https://www.baader-planetarium.com/en/},
\url{https://www.formatt-hitech.com/},
\url{https://www.fotga.com/},
\url{https://www.haidaphoto.com/en/},
\url{http://www.heliopan.de/},
\url{https://hoyafilter.com/},
\url{https://kenkoglobal.com/},
\url{https://kolarivision.com/},
\url{https://schneiderkreuznach.com/en/photo-optics},
\url{https://tiffen.com/},
\url{https://www.uvroptics.com/},
\url{https://www.zomei.com/}.
}

\examples{
photography_filters

}
\keyword{datasets}
