% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screens.R
\docType{data}
\name{arrigoni_screens}
\alias{arrigoni_screens}
\alias{criadolopez_screens}
\alias{howitec_screens}
\alias{huachangyarns_screens}
\alias{mallastextiles_screens}
\alias{oerlemansplastics_screens}
\alias{svensson_screens}
\title{Indexing of Climate Screens by Supplier}
\format{
A vector of character strings.

An object of class \code{character} of length 17.

An object of class \code{character} of length 40.

An object of class \code{character} of length 15.

An object of class \code{character} of length 45.

An object of class \code{character} of length 3.

An object of class \code{character} of length 32.
}
\usage{
arrigoni_screens

criadolopez_screens

howitec_screens

huachangyarns_screens

mallastextiles_screens

oerlemansplastics_screens

svensson_screens
}
\description{
Vectors of names of member objects containing spectral transmittance data for
climate, shade and protection screens used in agriculture grouped by
supplier. These can be used to extract subsets of spectra from the collection
stored in the \code{filter_mspct} object named \code{screens.mspct}.
}
\examples{
arrigoni_screens

# select filters supplied by Arrigoni
screens.mspct[arrigoni_screens[1:5]]

}
\references{
Catálogo de productos. Mallastextiles, Chimalhuacán, Edo. de México. \url{https://mallastextilesmx.com/}

"Samples folder". Howitec Netting.Bolsward, The Netherlands. \url{https://www.howitec.com/}

"Samples". Arrigoni Spa. Uggiate Trevano, Italia. \url{https://www.arrigoni.it}

"Samples". Criado & Lopez. El Ejido (Almería) España. \url{https://www.criadoylopez.com}

"Samples". oerlemans plastics BV. Genderen, The Netherlands. \url{https://www.oerlemansplastics.nl}

"samples". Jiangsu Huachang Yarns & Fabrics Co. Ltd. Jurong, Jiangsu, China. \url{https://www.hcpco.com}

Samples folder. Kinna, Sweden. \url{https://www.ludvigsvensson.com/}
}
\seealso{
\code{\link{screens.mspct}}
}
\keyword{datasets}
