% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theatrical-gels.r
\docType{data}
\name{theatrical_gels}
\alias{theatrical_gels}
\alias{lee_filters}
\alias{lee_gels}
\alias{rosco_filters}
\alias{rosco_gels}
\title{Indexing of theatrical "gels"}
\format{
A vector of character strings.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
theatrical_gels

lee_filters

lee_gels

rosco_filters

rosco_gels
}
\description{
Vectors of names of member objects with spectral transmittance for different
filters used on light sources in theatres, photo and video studios as
spectral modifiers. Frequently called "theatrical gels".
}
\note{
Instruments used: Shimadzu UV-2501 PC (Shimadzu Suzhou Instruments
  Manufacturing Co. Ltd, China) equipped with an integrating sphere for Rosco
  filters except no. 299 which was measured with an Agilent 8453 array
  spectrophotometer. Spectra for Lee gels have been digitized from
  the spectral plots in the Lee filters catalogue.
}
\examples{
theatrical_gels

}
\references{
\url{https://leefilters.com/lighting/colour-effect-lighting-filters/},
\url{https://emea.rosco.com/en/products/family/filters-and-diffusions}
}
\seealso{
\code{\link{filters.mspct}}
}
\keyword{datasets}
